Swap the old user model for a new one
This commit implements a custom user-model. The model extends AbstractUser, and thus inherits most of the functionality of the default user. The difference is now that we can add arbitrary fields to it.

Much of the touched files here represent simple imports of the default auth_user.User which has been changed.

Note that this commit requires a full database rebuild.