Update ShiftSlotGroup and ShiftSLot
Add meet_time and start_time to ShiftSlotGroup. Then change
'start' and 'end' fields on ShiftSlot to be TimeFields instead of
DateTimeFields, as the day is inferred from start_time on
ShiftSlotGroup.

The point here is that a ShiftSlot can have a time varying from slot to
slot in a ShiftSlotGroup, so we don't demand a specific time range in
the ShiftSlotGroup model. Instead we set a common meet_time and
start_time.