Moved titles for quotes outside quote-containers
The title for approved and unapproved quotes
were being positioned using grid-container items and gaps properties. In order to simplify it the titles have been moved outside the quote-containers.
There was also a bug from a misplaced <div> element outside of the if-statement resulting in a different rendering if there were unapproved quotes
pending.