Added custom manager to return quotes this semester
We needed to be able to render valid quotes of the semester meaning returning only the quotes that were created this semester. In addition to adding this custom manager to render is on the backend I also added a test to check whether or not the method returns the right quotes. The test has to override the auto_now=True attribute in the created_at field in order to test a non-valid timestamp.