Corrected rendering of correct semesters in archive
The different sub-sites of quotes did not properly render quotes or its corresponding semester. The archive now renders the last 15 semesters and clicking on one of them calls on the view rendering all quotes from that semester. Also fixed misc. in terms of renaming page titles from 'Arkiv' to 'V19' for the semester Spring 19 and so on