Resolve merge conflicts with commit in develop #9d827a922ad17dbc89cc41e11795884d98ac2d41
The conflicting commit rewrote the Quote as a timestamped model and changed
the way the manager was handled which was not supported by the current state
of this branch. The files in question then where checked out from develop
in order to make required changes for the build to start working again.