Bootstrap quote highscore view and presentation and fix button on pending
Initiates the view for rendering the quotes with the highest score.
Renders the first 10 quotes with the highest score from a given
semester string, and then another 10 rendered by all time.
In addition the view returns a a combined zip list which can
be used in the future so the table rendered in the template
can be styled as pairs of quotes, altough it needs handling
for when the lists are uneven which might happen at the beginning
of a new semester.