refactor: remove purchase model and deprecate parts of API
Somewhat breaking commit seeing as it attempts to remove the
Purchase model and move over some of the fields to the
ProductOrder model. All references to Purchase has been replaced
by ProductOrder where its applicable. Also deleting parts of the
rest API and tests not immediately compatible with this change.

Summary
* delete Purchase model
* move Purchase.source and Purchase.session field to ProductOrder model.
* refactor Socisession product properties to reflect this change
* delete SociBankAccountChargeView and related tests
* remove Purchase reference from admin view
* remove amount and purchase reference from ProductOrder