feat(admissions): add interview related resolvers and mutations
Implement resolvers and mutations related to the applicant interview
booking workflow. With these mutations and resolvers a user can
now request interviews available for booking in a way which is
easy to understand and interact with.

The implementation allows an applicant to book their interview
from all available interview times and can also request interview
times further ahead in time if none of the available options work
for the applicant.

Lastly rework implementation of email sending. Instead of sending
an email for each newly created applicant instance we aggregate
a list of emails of new applicants and mass mail them using bcc.
The applicant can then request a login token so we do not hit
any rate or spam limits on email