feat: model changes and queries for InternalGroup
Made some queries to fetch data in frontend, added "groupIcon".
Also added some Enums with reference to InternalGroup.Type.
I did notice that "Type" may be an unfortunate name for a field,
as I have to use uppercase in order to not get a conflict error
from graphene. Could be that I've overlooked something though.