fix: broken initial migrations
For some retarded reason all tests would fail because the whole
application could not manage to run all the migrations from zero
which it does during the test environment. When adding a new field
to the SociSession model (any field whatsoever) it would simply
halt at migration '0002' and state that the column (in this case
the new field) does not exist.

Adding a new field to another model (say users.User) did not screw
up anything, and getting rid of the breaking migration fixed
everything as well. The fix somehow was to squash all the economy
migrations and then re-create the migration for the new field.

I can't even. I hope no one ever will read this. I'm done.