feat(deps): upgrade Django to 4.2.3
drf-yasg2 kinda broke the entire upgrade. Ideally we should
re-document the relevant API-views, but managing to upgrade
the Django version has more important. With the upgrade
there were some issues with some of the asgi things that now
ships with django I think. This forced some upgrades of channels
and redis